+++ ph-schematron/src/main/java/com/helger/schematron/config/XPathConfigs.java
@@ -1,0 +1,15 @@
+package com.helger.schematron.config;
+
+import com.helger.schematron.pure.bound.xpath.PSXPathBoundSchema;
+import com.helger.xml.xpath.XPathHelper;
+
+import javax.xml.xpath.XPathFactory;
+import javax.xml.xpath.XPathFunctionResolver;
+import javax.xml.xpath.XPathVariableResolver;
+
+public class XPathConfigs {
+
+    private static final XPathFactory SAXON_FIRST = XPathHelper.createXPathFactorySaxonFirst ();
+
+    public static final XPathConfig DEFAULT = new XPathConfigImpl(SAXON_FIRST, null, null);
+}
