+++ ph-schematron/src/main/java/com/helger/schematron/pure/bound/PSBoundSchemaCacheKey.java
@@ -23,6 +23,7 @@ import javax.xml.xpath.XPathFunctionResolver;
 import javax.xml.xpath.XPathVariableResolver;
 
 import com.helger.schematron.config.XPathConfig;
+import com.helger.schematron.config.XPathConfigs;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.xml.sax.EntityResolver;
